+++ src/pytest_mysql/factories/noprocess.py
@@ -1,0 +1,66 @@
+# Copyright (C) 2013 by Clearcode <http://clearcode.cc>
+# and associates (see AUTHORS).
+
+# This file is part of pytest-mysql.
+
+# pytest-mysql is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Lesser General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+
+# pytest-mysql is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Lesser General Public License for more details.
+
+# You should have received a copy of the GNU Lesser General Public License
+# along with pytest-mysql.  If not, see <http://www.gnu.org/licenses/>.
+"""Process fixture factory for MySQL database."""
+
+import os
+from warnings import warn
+
+import pytest
+from _pytest.fixtures import FixtureRequest
+
+from pytest_mysql.config import get_config
+from pytest_mysql.executor_noop import NoopMySQLExecutor
+from pytest_mysql.port import get_port
+
+
+def mysql_noproc(host=None, port=None):
+    """
+    Process fixture factory for MySQL server.
+
+    :param str host: hostname
+    :param int port:
+    :rtype: func
+    :returns: function which makes a redis process
+
+    """
+
+    @pytest.fixture(scope="session")
+    def mysql_noproc_fixture(request: FixtureRequest):
+        """
+        Process fixture for MySQL server.
+
+        #. Get config.
+
+        :param request: fixture request object
+        :rtype: pytest_dbfixtures.executors.TCPExecutor
+        :returns: tcp executor
+
+        """
+        config = get_config(request)
+        mysql_port = int(port or config["port"] or 3306)
+        mysql_host = host or config["host"]
+
+        mysql_executor = NoopMySQLExecutor(
+            user=config["user"],
+            host=mysql_host,
+            port=mysql_port,
+        )
+        with mysql_executor:
+            yield mysql_executor
+
+    return mysql_noproc_fixture